fix: rename package from database to utils for MovioDataBase
