+++ app/src/main/java/com/cairosquad/movio/di/localDataSourceModule.kt
@@ -1,13 +1,13 @@
 package com.cairosquad.movio.di
 
 import androidx.room.Room
-import com.cairosquad.local.database.MovioDataBase
 import com.cairosquad.local.search.cache.cacheDataSourceImpl
 import com.cairosquad.local.search.cache.dao.CacheDao
 import com.cairosquad.local.search.discovery.DiscoveryDataSourceImpl
 import com.cairosquad.local.search.discovery.dao.DiscoveryDao
 import com.cairosquad.local.search.recent.LocalRecentSearchDataSourceImpl
 import com.cairosquad.local.search.recent.dao.LocalRecentSearchDao
+import com.cairosquad.local.utils.MovioDataBase
 import com.cairosquad.repository.search.data_source.local.CacheDataSource
 import com.cairosquad.repository.search.data_source.local.DiscoveryDataSource
 import com.cairosquad.repository.search.data_source.local.LocalRecentSearchDataSource
